+++ CHANGELOG.md
@@ -10,21 +10,51 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 
 ## Unreleased
 
+### Fixed
+
+- Incorrect column title on `code42 trusted-activities bulk create` command help text.
+
+### Added
+
+- New bulk commands to manage user roles
+    - `code42 users bulk add-roles`
+    - `code42 users bulk remove-roles`
+
+- New option `--include-roles` on `code42 users list` that includes the roles for all users.
+
+- New command `code42 users show <username>` that prints all the details of that user.
+
+- New commands to view orgs
+    - `code42 users orgs list`
+    - `code42 users orgs show <org-uid>`
+
+## 1.10.0 - 2021-10-05
+
 ### Added
 
 - New option `--include-legal-hold-membership` on command `code42 users list` that includes the legal hold matter name and ID for any user on legal hold.
 
 - New commands for deactivating/reactivating Code42 user accounts:
-  - `code42 users deactivate`
-  - `code42 users reactivate`
-  - `code42 users bulk deactivate`
-  - `code42 users bulk reactivate`
+    - `code42 users deactivate`
+    - `code42 users reactivate`
+    - `code42 users bulk deactivate`
+    - `code42 users bulk reactivate`
 
 - `code42 profile use` now prompts you to select a profile when not given a profile name argument.
+
+- New `trusted-activities` commands for managing trusted activities and resources:
+    - `code42 trusted-activities create` to create a trusted activity.
+    - `code42 trusted-activities update` to update a trusted activity.
+    - `code42 trusted-activities remove` to remove a trusted activity.
+    - `code42 trusted-activities list` to print the details of all trusted activities.
+    - `code42 trusted-activities bulk create` to bulk create trusted activities from a CSV file.
+    - `code42 trusted-activities bulk update` to bulk update trusted activities from a CSV file.
+    - `code42 trusted-activities bulk remove` to bulk remove trusted activities from a CSV file.
 
 ### Fixed
 
 - Bug where `audit-logs search` with `--use-checkpoint` option was causing output formatting problems.
+- Improve error message for `code42 users list`, `code42 devices list`, `code42 devices list-backup-sets`
 
 ## 1.9.0 - 2021-08-19